pyproject.toml
[project]
name = "architect"
version = "0.1.0"
description = "The Architect — Sovereign Development & Autonomy Platform"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "redis[hiredis]>=5.0.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
]

[project.scripts]
architect = "architect.cli:app"
